requirements.txt
flask==2.0.3
pymysql==1.0.2
cryptography==38.0.1
werkzeug==2.2.2
mysql-connector-python
cloudpathlib[s3]